+++ src/map.c
@@ -3,6 +3,8 @@
 #include <Rversion.h>
 #include <Rinternals.h>
 #include "coerce.h"
+#include "conditions.h"
+#include "utils.h"
 
 void copy_names(SEXP from, SEXP to) {
   if (Rf_length(from) != Rf_length(to))
@@ -16,15 +18,11 @@ void copy_names(SEXP from, SEXP to) {
 }
 
 void check_vector(SEXP x, const char *name) {
-  if (Rf_isNull(x) || Rf_isVector(x) || Rf_isPairList(x))
+  if (Rf_isNull(x) || Rf_isVector(x) || Rf_isPairList(x)) {
     return;
+  }
 
-  Rf_errorcall(
-    R_NilValue,
-    "`%s` is not a vector (%s)",
-    name,
-    Rf_type2char(TYPEOF(x))
-  );
+  stop_bad_type(x, "a vector", NULL, name);
 }
 
 // call must involve i
@@ -46,8 +44,10 @@ SEXP call_loop(SEXP env, SEXP call, int n, SEXPTYPE type, int force_args) {
 #else
     SEXP res = PROTECT(Rf_eval(call, env));
 #endif
-    if (type != VECSXP && Rf_length(res) != 1)
-      Rf_errorcall(R_NilValue, "Result %i is not a length 1 atomic vector", i + 1);
+    if (type != VECSXP && Rf_length(res) != 1) {
+      SEXP ptype = PROTECT(Rf_allocVector(type, 0));
+      stop_bad_element_vector(res, i + 1, ptype, 1, "Result", NULL, false);
+    }
 
     set_vector_value(out, i, res, 0);
     UNPROTECT(1);
@@ -110,7 +110,12 @@ SEXP map2_impl(SEXP env, SEXP x_name_, SEXP y_name_, SEXP f_name_, SEXP type_) {
     return Rf_allocVector(type, 0);
   }
   if (nx != ny && !(nx == 1 || ny == 1)) {
-    Rf_errorcall(R_NilValue, "`.x` (%i) and `.y` (%i) are different lengths", nx, ny);
+    Rf_errorcall(R_NilValue,
+                 "Mapped vectors must have consistent lengths:\n"
+                 "* `.x` has length %d\n"
+                 "* `.y` has length %d",
+                 nx,
+                 ny);
   }
   int n = (nx > ny) ? nx : ny;
 
@@ -133,8 +138,9 @@ SEXP pmap_impl(SEXP env, SEXP l_name_, SEXP f_name_, SEXP type_) {
   SEXP l_val = PROTECT(Rf_eval(l, env));
   SEXPTYPE type = Rf_str2type(CHAR(Rf_asChar(type_)));
 
-  if (!Rf_isVectorList(l_val))
-    Rf_errorcall(R_NilValue, "`.x` is not a list (%s)", Rf_type2char(TYPEOF(l_val)));
+  if (!Rf_isVectorList(l_val)) {
+    stop_bad_type(l_val, "a list", NULL, l_name);
+  }
 
   // Check all elements are lists and find maximum length
   int m = Rf_length(l_val);
@@ -143,7 +149,7 @@ SEXP pmap_impl(SEXP env, SEXP l_name_, SEXP f_name_, SEXP type_) {
     SEXP j_val = VECTOR_ELT(l_val, j);
 
     if (!Rf_isVector(j_val) && !Rf_isNull(j_val)) {
-      Rf_errorcall(R_NilValue, "Element %i is not a vector (%s)", j + 1, Rf_type2char(TYPEOF(j_val)));
+      stop_bad_element_type(j_val, j + 1, "a vector", NULL, l_name);
     }
 
     int nj = Rf_length(j_val);
@@ -162,8 +168,9 @@ SEXP pmap_impl(SEXP env, SEXP l_name_, SEXP f_name_, SEXP type_) {
     SEXP j_val = VECTOR_ELT(l_val, j);
     int nj = Rf_length(j_val);
 
-    if (nj != 1 && nj != n)
-      Rf_errorcall(R_NilValue, "Element %i has length %i, not 1 or %i.", j + 1, nj, n);
+    if (nj != 1 && nj != n) {
+      stop_bad_element_length(j_val, j + 1, n, NULL, ".l", true);
+    }
   }
 
   SEXP l_names = PROTECT(Rf_getAttrib(l_val, R_NamesSymbol));